refactor: 서비스 목업 코드 수정 #16
